requirements.txt
Flask==2.3.2
Flask-SQLAlchemy
scikit-learn
pandas
numpy
joblib
matplotlib
seaborn
imbalanced-learn
openai
google-generativeai
python-dotenv